refactor: setup-related functionality into separate module
